changed directory structure of helpers module
