+++ src/civitconfig/__main__.py
@@ -1,12 +1,9 @@
 #!/usr/bin/env python3
 
-import sys
 import traceback
-from operator import itemgetter
 
-from helpers.styler import Styler
-from helpers.exceptions import UnexpectedException
-from helpers.utils import set_verbose, run_verbose, print_verbose, print_exc, DefaultOptions, sprint
+from helpers.core.utils import UnexpectedException, set_verbose, run_verbose, print_verbose, print_exc, sprint
+from helpers.options import DefaultOptions
 from civitconfig.args.argparser import get_args
 from civitconfig.data.configmanager import ConfigManager
