+++ src/civitconfig/data/config/config.py
@@ -6,10 +6,8 @@ import shutil
 from typing import Dict, List, Union
 
 from helpers.sorter import basic, tags
-from helpers.exceptions import UnexpectedException
-from helpers.styler import Styler
-from helpers.utils import getDate, sprint
-from helpers.validation import Validation
+from helpers.core.utils import Styler, UnexpectedException, getDate, sprint
+from helpers.core._validation import Validation
 
 DEFAULT_CONFIG = {
     "version": "1",
