+++ src/civitconfig/data/config/defaultconfig.py
@@ -1,5 +1,5 @@
-from helpers.exceptions import InputException, UnexpectedException
-from helpers.utils import BatchOptions, DefaultOptions
+from helpers.core.utils import InputException, UnexpectedException
+from helpers.options import DefaultOptions
 from .config import Config
