+++ src/civitconfig/data/config/sorterconfig.py
@@ -1,8 +1,7 @@
 import os
 import shutil
 import sys
-from helpers.exceptions import InputException, ResourcesException
-from helpers.utils import getDate, print_exc, sprint
+from helpers.core.utils import InputException, ResourcesException, getDate, print_exc, sprint
 from helpers.sorter.utils import import_sort_model
 from .config import Config
