+++ src/civitdl/__main__.py
@@ -1,12 +1,9 @@
 import traceback
-from operator import itemgetter
 
 from .batch.batch_download import batch_download, BatchOptions
 from .args.argparser import get_args
 
-from helpers.exceptions import UnexpectedException
-from helpers.sourcemanager import SourceManager
-from helpers.utils import print_verbose, run_verbose, print_exc, set_verbose, sprint
+from helpers.core.utils import print_verbose, run_verbose, print_exc, set_verbose, sprint
 
 
 def main():
