+++ src/civitdl/api/sorter.py
@@ -1,4 +1,4 @@
 import re
 
-from helpers.constants import BLACKLISTED_DIR_CHARS
+from helpers.core.constants import BLACKLISTED_DIR_CHARS
 from helpers.sorter.utils import DirName, SorterData
