--- src/helpers/constants.py
@@ -1,1 +1,0 @@
-BLACKLISTED_DIR_CHARS = ['<', '>', ':', '"', '/', '\\', '|', '?', '*']
