+++ src/helpers/core/constants.py
@@ -1,3 +1,5 @@
 from appdirs import AppDirs
 
 app_dirs = AppDirs('civitdl', 'Owen Truong')
+
+BLACKLISTED_DIR_CHARS = ['<', '>', ':', '"', '/', '\\', '|', '?', '*']
