--- src/helpers/dep.txt
@@ -1,17 +1,0 @@
-Main Directories
-- adep < bdep < cdep < ddep < edep
-- files in adep may not access other deps's files.
-- files edep may access all other deps's public files.
-- files in the same directory level may not access each other unless one is private and the other is public (public can access private files)
-- files may access public dep files in subdirectory.
-
-Subdirectories
-- At most 2 level deep of subdirectories per main directory.
-- main < sub1 < sub2
-
-File functions
-- public functions may access private functions in the same file.
-- There may be only 4 levels of dependencies between functions in each file.
-
-
-
