+++ src/helpers/sourcemanager.py
@@ -1,18 +1,9 @@
 from dataclasses import dataclass
-from datetime import datetime
-import json
 import os
 import re
-import sys
-from typing import Callable, Dict, Iterable, List, Literal, Union
-import importlib.util
-import concurrent.futures
-import requests
-from tqdm import tqdm
+from typing import List, Literal, Union
 
-from helpers.styler import Styler
-from helpers.exceptions import CustomException, InputException, UnexpectedException
-from helpers.utils import print_verbose
+from helpers.core.utils import print_verbose, InputException, UnexpectedException
 
 
 @dataclass
